fix: Dockerfile
reason: to build a docker image is necessary to have a .jar file of the application, but while trying to start the mvn build package command, a test of our spring application was performed and because spring needs a running database, the test failed and was unable to generate the .jar file.
Reason: To build a Docker image, it's necessary to have a .jar file of the application. During the initial build, running the mvn clean package command failed because the tests required a running database, which was not available. This caused the build to fail and prevented the .jar file from being generated.

Solution: This issue was resolved by adding the -DskipTests flag to the mvn package command. To avoid issues for users who may not have Maven installed, we updated the Dockerfile to use a Maven Docker image for the build process. This ensures that the .jar file is generated correctly, regardless of whether Maven is installed on the host system.